requirements.txt
django==4.2
django-filter==23.1
django-import-export==3.1.0
djangorestframework==3.14.0
djoser==2.1.0
requests==2.28.2
pytz==2023.3
